NIFI-11035 Replaced remaining JUnit 4 assertions in nifi-commons with JUnit 5
- Replaced Groovy asserts with JUnit 5 assertions and Groovy shouldFail method Junit 5 with assertThrow method

This closes #6880

+++ nifi-commons/nifi-utils/src/test/java/org/apache/nifi/util/timebuffer/TestTimedBuffer.java
@@ -20,8 +20,8 @@ import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.TimeUnit;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 public class TestTimedBuffer {
 
@@ -90,7 +90,7 @@ public class TestTimedBuffer {
                 return oldValue;
             }
 
-            return new TimestampedLong(oldValue.getValue().longValue() + toAdd.getValue().longValue());
+            return new TimestampedLong(oldValue.getValue() + toAdd.getValue());
         }
 
         @Override
